Move LocalStackAwsV1ClientHelper into localstack-test
