+++ java/cdk/src/test/java/sleeper/cdk/testutils/LocalStackTestBase.java
@@ -37,7 +37,7 @@ import sleeper.localstack.test.SleeperLocalStackContainer;
 import java.util.List;
 
 import static java.util.stream.Collectors.toUnmodifiableList;
-import static sleeper.configuration.testutils.LocalStackAwsV1ClientHelper.buildAwsV1Client;
+import static sleeper.localstack.test.LocalStackAwsV1ClientHelper.buildAwsV1Client;
 
 @Testcontainers
 public abstract class LocalStackTestBase {
