+++ java/common/common-invoke-tables/src/test/java/sleeper/invoke/tables/InvokeForTablesIT.java
@@ -41,8 +41,8 @@ import java.util.stream.Stream;
 import static java.util.stream.Collectors.toUnmodifiableList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static sleeper.configuration.testutils.LocalStackAwsV1ClientHelper.buildAwsV1Client;
 import static sleeper.core.table.TableStatusTestHelper.uniqueIdAndName;
+import static sleeper.localstack.test.LocalStackAwsV1ClientHelper.buildAwsV1Client;
 
 @Testcontainers
 public class InvokeForTablesIT {
