+++ java/common/common-job/src/test/java/sleeper/job/common/action/ChangeMessageVisibilityTimeoutActionIT.java
@@ -33,7 +33,7 @@ import java.util.Map;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.configuration.testutils.LocalStackAwsV1ClientHelper.buildAwsV1Client;
+import static sleeper.localstack.test.LocalStackAwsV1ClientHelper.buildAwsV1Client;
 
 @Testcontainers
 public class ChangeMessageVisibilityTimeoutActionIT {
