+++ java/common/common-task/src/test/java/sleeper/task/common/QueueMessageCountIT.java
@@ -28,7 +28,7 @@ import sleeper.localstack.test.SleeperLocalStackContainer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static sleeper.configuration.testutils.LocalStackAwsV1ClientHelper.buildAwsV1Client;
+import static sleeper.localstack.test.LocalStackAwsV1ClientHelper.buildAwsV1Client;
 
 @Testcontainers
 class QueueMessageCountIT {
