+++ java/ingest/ingest-batcher-submitter/src/test/java/sleeper/ingest/batcher/submitter/IngestBatcherSubmitterLambdaIT.java
@@ -40,8 +40,8 @@ import sleeper.localstack.test.SleeperLocalStackContainer;
 import java.time.Instant;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.configuration.testutils.LocalStackAwsV1ClientHelper.buildAwsV1Client;
 import static sleeper.core.properties.testutils.InstancePropertiesTestHelper.createTestInstanceProperties;
+import static sleeper.localstack.test.LocalStackAwsV1ClientHelper.buildAwsV1Client;
 
 @Testcontainers
 public class IngestBatcherSubmitterLambdaIT {
