+++ java/ingest/ingest-runner/src/test/java/sleeper/ingest/runner/IngestRecordsLocalStackITBase.java
@@ -28,7 +28,7 @@ import sleeper.localstack.test.SleeperLocalStackContainer;
 import sleeper.statestore.s3.S3StateStore;
 import sleeper.statestore.s3.S3StateStoreCreator;
 
-import static sleeper.configuration.testutils.LocalStackAwsV1ClientHelper.buildAwsV1Client;
+import static sleeper.localstack.test.LocalStackAwsV1ClientHelper.buildAwsV1Client;
 import static sleeper.parquet.utils.HadoopConfigurationLocalStackUtils.getHadoopConfiguration;
 
 @Testcontainers
